Criei um novo ficheiro/classe para as funções que estavam na main e adicionei alguns comments que faltavam
